fix(cs): 修复发送文件时文件头部多出4字节的bug

+++ cs/hsocket/hsocket.cs
@@ -96,7 +96,7 @@ public class HTcpSocket : HSocket
     public void SendFile(FileStream file, string filename)
     {
         // get file size
-        long filesize = file.Length;
+        int filesize = (int)file.Length;
         // file header
         Send(Encoding.UTF8.GetBytes(filename));  // filename
         Send(new byte[1] { 0 });  // name end
